requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
watchdog==3.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pywebview==5.4
bcrypt==4.0.1 